Create initial version of LabTest pattern
The new LabTask class handles doing research in a lab as part of a Task.

Changes:
- Create a new abstract LabTask class that is a subclass of a Task
- LabTask provides management of the Lab addign & removing researchers
- It selected the most appropriate lab for the target Scientific Study
- There is a central execute method to perform research in a lab on a study
- PerformLaboratoryResearch uses the new LabTask
- PerformLaboratoryResearch also uses the new ExperienceImpact via a factory method
- The Task classes relevent to Scientifc Study are moved to the new core.science.task package
- Create new LabTask unit test that tests the lab tasks

Part of #1286
