+++ mars-sim-core/src/main/java/com/mars_sim/core/person/ai/task/ObserveAstronomicalObjects.java
@@ -21,6 +21,7 @@ import com.mars_sim.core.person.ai.task.util.Task;
 import com.mars_sim.core.person.ai.task.util.TaskPhase;
 import com.mars_sim.core.science.ScienceType;
 import com.mars_sim.core.science.ScientificStudy;
+import com.mars_sim.core.science.task.ResearchScientificStudy;
 import com.mars_sim.core.structure.Settlement;
 import com.mars_sim.core.structure.building.function.AstronomicalObservation;
 import com.mars_sim.core.structure.building.function.FunctionType;
