+++ mars-sim-core/src/main/java/com/mars_sim/core/person/ai/task/StudyFieldSamples.java
@@ -30,6 +30,7 @@ import com.mars_sim.core.person.ai.task.util.TaskPhase;
 import com.mars_sim.core.resource.ResourceUtil;
 import com.mars_sim.core.science.ScienceType;
 import com.mars_sim.core.science.ScientificStudy;
+import com.mars_sim.core.science.task.ResearchScientificStudy;
 import com.mars_sim.core.structure.Lab;
 import com.mars_sim.core.structure.Settlement;
 import com.mars_sim.core.structure.building.Building;
