+++ mars-sim-core/src/main/java/com/mars_sim/core/science/task/PerformLaboratoryResearchMeta.java
@@ -4,7 +4,7 @@
  * @Date 2021-10-05
  * @author Scott Davis
  */
-package com.mars_sim.core.person.ai.task.meta;
+package com.mars_sim.core.science.task;
 
 import java.util.HashSet;
 import java.util.List;
@@ -16,7 +16,6 @@ import com.mars_sim.core.person.Person;
 import com.mars_sim.core.person.ai.fav.FavoriteType;
 import com.mars_sim.core.person.ai.job.util.JobType;
 import com.mars_sim.core.person.ai.role.RoleType;
-import com.mars_sim.core.person.ai.task.PerformLaboratoryResearch;
 import com.mars_sim.core.person.ai.task.util.FactoryMetaTask;
 import com.mars_sim.core.person.ai.task.util.Task;
 import com.mars_sim.core.person.ai.task.util.TaskJob;
@@ -53,7 +52,7 @@ public class PerformLaboratoryResearchMeta extends FactoryMetaTask {
 
     @Override
     public Task constructInstance(Person person) {
-        return new PerformLaboratoryResearch(person);
+        return PerformLaboratoryResearch.createTask(person);
     }
 
     /**
@@ -106,12 +105,12 @@ public class PerformLaboratoryResearchMeta extends FactoryMetaTask {
      */
     private double getStudyScore(Person person, double base, ScienceType science) {
         double score = 0D;
-        Lab lab = PerformLaboratoryResearch.getLocalLab(person, science);
+        Lab lab = LabTask.getLocalLab(person, science);
         if (lab != null) {
             score = base;
 
             // Get lab building crowding modifier.
-            score *= PerformLaboratoryResearch.getLabCrowdingModifier(person, lab);
+            score *= LabTask.getLabCrowdingModifier(person, lab);
 
             // If researcher's current job isn't related to study science, divide by two.
             JobType job = person.getMind().getJob();
